backend/requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# ML & Computer Vision
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.8

# Speech Processing
openai-whisper>=20231117
gtts>=2.4.0

# Data Processing
pandas>=2.1.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
